requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.4.1
Flask-Login==0.2.7
Flask-SQLAlchemy==1.0
Flask-Script==0.6.6
Flask-WTF==0.9.4
Jinja2==2.7.3
bbcode==1.0.21
python-ldap==2.4.19
